requirements.txt
altgraph==0.17.4
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
auth==0.5.3
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorclass==2.2.2
confusable-homoglyphs==3.3.1
cryptography==44.0.0
defusedxml==0.7.1
delocate==0.12.0
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.1
django-comuni-italiani==1.4.1
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.3
django-registration==5.1.0
django-registration-redux==2.13
django-restframework==0.0.1
django-sslserver==0.22
django-vite==3.0.5
django-webpack-loader==3.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
docopt==0.6.2
eventlet==0.38.2
falcon==4.0.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
macholib==1.16.3
MarkupSafe==3.0.2
mccabe==0.7.0
mongoengine==0.29.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
pip-upgrader==1.4.15
platformdirs==4.3.6
prompt_toolkit==3.0.48
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pymongo==4.10.1
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2022.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
terminaltables==3.1.10
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
